Fix build system to avoid installing minified code as source.
When building with a non-empty object directory, the dependency on a source
file could be resolved to the minifier output if it had the same name.

Now, the minifier output gets a distinct name which is corrected on
installation. The (small) price to pay is that the minified file name given
in the generated source map is incorrect, but that field is optional and
its usefulness debatable.

On the upside it is now possible to build in the source tree, so the check
that forced creating an object directory first was removed.