Overhaul build system.
ES6 modules use the name of the source file in "import" statements and
renaming source files in the build process breaks this.

It's now the responsibility of the user to ensure that source files have a
different name from the output files. The default scheme is <prog>-src.js
for <prog>.js.

The advantages are a much simpler build system (no file copying) and in
projects that consist of several files, not all files need to have silly
"-src" names.